+++ .github/workflows/build_and_deploy_docs.yml
@@ -28,6 +28,7 @@ jobs:
       with:
         fetch-tags: true
         fetch-depth: 0
+        lfs: true
     - uses: prefix-dev/setup-pixi@v0.8.1
       with:
         pixi-version: v0.53.0
