+++ .github/workflows/build_docs.yml
@@ -14,6 +14,7 @@ jobs:
       with:
         fetch-tags: true
         fetch-depth: 0
+        lfs: true
     - uses: prefix-dev/setup-pixi@v0.8.1
       with:
         pixi-version: v0.53.0
