+++ .github/workflows/pixi_test.yml
@@ -24,6 +24,8 @@ jobs:
     steps:
     - name: Checkout repository
       uses: actions/checkout@v4
+      with:
+        lfs: true
     - uses: prefix-dev/setup-pixi@v0.8.1
       with:
         pixi-version: v0.53.0
