+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 Best viewed [here](https://qstheory.github.io/fftarray/main/changelog.html).
 
 ## Unreleased
-
 - The `space`, `eager` and `factors_applied` arguments of different functions now also accept dictionaries mapping from the name of the dimension to the value.
+- Add the two species example as a notebook to the documentation.
 
 ## fftarray 0.5.0 (17 July 2025)
 - Initial public release.
