+++ docs/source/_static/thumbmails/examples/two_species_groundstate.jpg
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9386cc4cb9f4c6ce62fea61a2ff19f413f19855c4727cae62a39f8bcdd108eae
+size 30389
