+++ examples/.gitattributes
@@ -1,0 +1,1 @@
+two_species_groundstate.ipynb filter=lfs diff=lfs merge=lfs -text
