+++ pixi.lock
@@ -5495,8 +5495,8 @@ packages:
   timestamp: 1734873425804
 - pypi: ./
   name: fftarray
-  version: 0.5.1.dev16+g3e66cbdcc
-  sha256: 7dd170d8a70be6aa553402ea4d3934f0797103e97c95fffcab37980eafe61b6f
+  version: 0.5.1.dev11+g45d121731.d20250902
+  sha256: aa21ecbc5b4e3927aef002624990f9ba244848ec83c1eb17911931c8a7c11504
   requires_dist:
   - array-api-compat>=1.11.0
   - numpy>=2.0
