+++ tests/end_to_end/test_examples.py
@@ -3,15 +3,18 @@
 
 import io
 import types
+import os
 
 import pytest
 from IPython import get_ipython
 from nbformat import read
 from IPython.core.interactiveshell import InteractiveShell
 
+import jax
+
+
 from tests.helpers import XPS
 import fftarray as fa
-from examples.two_species_groundstate import calc_ground_state_two_species
 
 @pytest.mark.parametrize("xp", XPS)
 @pytest.mark.parametrize("nb_path", [
@@ -21,6 +24,16 @@ from examples.two_species_groundstate import calc_ground_state_two_species
     "Derivative",
 ])
 def test_notebooks(xp, nb_path):
+    run_notebook(xp=xp, nb_path=nb_path)
+
+
+def test_two_species_notebook():
+    """
+        The two-species example explicitly utilizes jax.
+    """
+    run_notebook(xp=jax.numpy, nb_path="two_species_groundstate")
+
+def run_notebook(xp, nb_path) -> None:
     # load the notebook object
     with io.open(f"examples/{nb_path}.ipynb", 'r', encoding='utf-8') as f:
         nb = read(f, 4)
@@ -37,6 +50,11 @@ def test_notebooks(xp, nb_path):
     save_user_ns = shell.user_ns
     shell.user_ns = mod.__dict__
 
+    os.environ["TEST_MODE"] = "True"
+    # Force the matplotlib backend to be non-interactive.
+    # Otherwise this test fails for example on a macOS notebook.
+    os.environ["MPLBACKEND"] = "agg"
+
     try:
         for cell in nb.cells:
             if cell.cell_type == 'code':
@@ -50,7 +68,3 @@ def test_notebooks(xp, nb_path):
                     exec(code, mod.__dict__)
     finally:
         shell.user_ns = save_user_ns
-
-def test_two_species_groundstate():
-
-    calc_ground_state_two_species(N_iter=3)
